Backend-feedback/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
attrs==25.3.0
blinker==1.8.2
cffi==1.17.1
click==8.1.8
cryptography==45.0.4
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.14
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.16.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.1.3
packaging==25.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.4.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.44.0
typing_extensions==4.13.2
uvicorn==0.33.0
Werkzeug==3.0.6
zipp==3.20.2
